backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyodbc==5.2.0
python-socketio==5.11.4